Move fetchRetry tests adjacent to implementation
